Add EFI stub to kernel build
This changes the kernel config to enable the EFI stub, which allows the
kernel to be signed and verified as a PE/COFF file and used directly in
the UEFI Secure Boot boot chain.

It also prevents future potential breakage when/if the xen-unified
binary is used on systems with the shim loader, because if Xen detects
that it has been loaded by shim, it will attempt to verify the dom0
kernel that has been embedded into it, which will fail if it does not
contain the EFI stub and hasn't been signed with a valid signature.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>