Disable build of i40iw driver
Since we have an updated out-of-tree version for the i40e driver module,
the i40iw driver from kernel 4.19 is no more compatible with it because
of missing symbols in the new i40e driver. depmod is issuing a warning
about those missing symbols when updating the kernel package.

This patch disables the compilation of the i40iw driver.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>